requirements.txt
tweepy
pandas
google
networkx
pymongo
dnspython
faker